move DynamicConfiguration package to correct service package.
