+++ javamesh-agentcore/javamesh-agentcore-core/src/main/java/com/huawei/apm/core/service/dynamicconfig/ConfigChangedEvent.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.huawei.apm.core.dynamicconfig;
+package com.huawei.apm.core.service.dynamicconfig;
 
 import java.util.EventObject;
 import java.util.Objects;
